backend/requirements.txt
# DataStatz Backend — Python dependencies

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Data processing — pinned to versions with pre-built wheels for Python 3.11
pandas==2.1.4
numpy==1.26.4
openpyxl==3.1.3
xlrd==2.0.1

# Stats & ML
scipy==1.11.4
scikit-learn==1.4.2
shap>=0.42.0

# Validation
pydantic==2.7.1

# Supabase (shared reports persistence)
supabase==2.4.3
python-dotenv==1.0.1
